docs(env): remove semicolons from example .env file
The semicolons in the .env file provided in the documentation are not
required. (The .env file is not a TypeScript or JavaScript file.)
Moreover, in the example provided, with the semicolon present, the value
for e.g. the DATABASE_USER environment variable will end up being the
string "test;", not "test". Presumably the latter is what is actually
intended.